Updated code to reflect number relative to 1.0 rather than 0.84

+++ PPMTool/Data/CapacityQueryItem.cs
@@ -14,9 +14,9 @@ namespace PPMTool.Data
 
         public DateTime EndDate { get; }
 
-        public int AvailabilityPercent { get; }
+        public double AvailabilityPercent { get; }
 
-        public CapacityQueryItem(Person person, DateTime startDate, DateTime endDate, int availabilityPercent)
+        public CapacityQueryItem(Person person, DateTime startDate, DateTime endDate, double availabilityPercent)
         {
             Person = person;
             StartDate = startDate;
